requirements.txt
python-dotenv
sqlalchemy
telethon